Fix - Disability Weight methodological error
FIX
---
Previously: Interventions affected the `combined_dw`, which reduced the
disability weight of asthma, after it had been blended with healthy
disability weights. This overestimates the effects of asthma by also
reducing the disability weight of simply being healthy.

The new implementation modifies the disability weight of asthma in
particular, before it is combined. This results in healthy life years
gained of about 1/3rd of the previous estimation.